eslint: fix react-refresh/only-export-components
