+++ src/components/playlist/playerErrors/Entry.jsx
@@ -132,4 +132,6 @@ class PlayerErrorsEntry extends Component {
     }
 }
 
-export default withNavigate(withSearchParams(PlayerErrorsEntry))
+PlayerErrorsEntry = withNavigate(withSearchParams(PlayerErrorsEntry))
+
+export default PlayerErrorsEntry
