+++ src/components/settings/users/Entry.jsx
@@ -5,12 +5,11 @@ import { CSSTransitionLazy } from 'components/adapted/ReactTransitionGroup'
 import ConfirmationBar from 'components/generics/ConfirmationBar'
 import ControlLink from 'components/generics/ControlLink'
 import Notification, { NotifiableForTable } from 'components/generics/Notification'
-import {
-    IsNotSelf,
-    IsUserManager,
-    permissionLevels
-} from 'components/permissions/Users'
+import { IsNotSelf, IsUserManager } from 'components/permissions/Users'
 import { userPropType } from 'serverPropTypes/users'
+
+import Marked from './Marked'
+import PermissionText from './PermissionText'
 
 export default class SettingsUsersEntry extends Component {
     static propTypes = {
@@ -111,33 +110,4 @@ export default class SettingsUsersEntry extends Component {
             </tr>
         )
     }
-}
-
-const PermissionText = ({level}) => {
-    if (!level) {
-        return null
-    }
-
-    const permissionText = permissionLevels[level]
-
-    return (
-        <span className="permission-text">
-            {permissionText.substring(0, 1)}
-            <span className="hideable">
-                {permissionText.substring(1)}
-            </span>
-        </span>
-    )
-
-}
-
-
-const Marked = ({marked}) => {
-    if (!marked) {
-        return null
-    }
-
-    return (
-        <i className="las la-check"></i>
-    )
 }
