+++ src/components/settings/users/Marked.jsx
@@ -1,0 +1,9 @@
+export default function Marked({ marked }) {
+    if (!marked) {
+        return null
+    }
+
+    return (
+        <i className="las la-check"></i>
+    )
+}
