+++ src/components/settings/users/PermissionText.jsx
@@ -1,0 +1,18 @@
+import { permissionLevels } from 'utils/permissions';
+
+export default function PermissionText({ level }) {
+    if (!level) {
+        return null
+    }
+
+    const permissionText = permissionLevels[level]
+
+    return (
+        <span className="permission-text">
+            {permissionText.substring(0, 1)}
+            <span className="hideable">
+                {permissionText.substring(1)}
+            </span>
+        </span>
+    )
+}
