+++ src/utils/permissions.js
@@ -1,0 +1,9 @@
+export const permissionLevels = {
+    u: 'user',
+    m: 'manager',
+    p: 'player'
+}
+
+export const mapStateToProps = (state) => ({
+    user: state.authenticatedUser
+})
